+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.4.0](https://github.com/dash-ui/mq/compare/v0.3.0...v0.4.0) (2020-07-27)
+
+### ⚠ BREAKING CHANGES
+
+- Is only compatible with @dash-ui/styles@>=0.8.0
+
+- rename all instances of "variables" to "tokens" ([dd8532a](https://github.com/dash-ui/mq/commit/dd8532a4aee8df7104616b3fbd22c378c97b12b5))
+
 ## [0.3.0](https://github.com/dash-ui/mq/compare/v0.2.0...v0.3.0) (2020-07-23)
 
 ### ⚠ BREAKING CHANGES